requirements.txt
Flask
python-dotenv
groq
paho-mqtt